* moved ComPtr to detail folder
* removed unused forward declaration
* removed unnecessary macro guard
