+++ include/orhi/impl/dx12/DescriptorSet.h
@@ -10,7 +10,7 @@
 #include <orhi/impl/dx12/BackendTraits.h>
 #include <orhi/impl/dx12/Device.h>
 
-#include <orhi/impl/dx12/ComPtr.h>
+#include <orhi/impl/dx12/detail/ComPtr.h>
 
 struct ID3D12DescriptorHeap;
