+++ include/orhi/impl/dx12/Device.h
@@ -14,7 +14,7 @@
 #include <memory>
 #include <vector>
 
-#include <orhi/impl/dx12/ComPtr.h>
+#include <orhi/impl/dx12/detail/ComPtr.h>
 
 struct ID3D12Device;
