+++ include/orhi/impl/dx12/Instance.h
@@ -8,7 +8,7 @@
 
 #include <orhi/api/TInstance.h>
 #include <orhi/impl/dx12/BackendTraits.h>
-#include <orhi/impl/dx12/ComPtr.h>
+#include <orhi/impl/dx12/detail/ComPtr.h>
 
 #include <memory>
