+++ include/orhi/impl/dx12/Queue.h
@@ -8,7 +8,7 @@
 
 #include <orhi/api/TQueue.h>
 #include <orhi/impl/dx12/BackendTraits.h>
-#include <orhi/impl/dx12/ComPtr.h>
+#include <orhi/impl/dx12/detail/ComPtr.h>
 
 struct ID3D12CommandQueue;
