+++ include/orhi/impl/dx12/detail/ComPtr.h
@@ -1,0 +1,10 @@
+/**
+* @project: orhi (OpenRHI)
+* @author: Jian Bang Xu
+* @licence: MIT
+*/
+
+#pragma once
+
+#define NOGDI
+#include <wrl.h>
